VMS-15486: add UT for group 'primaryStream' and 'secondaryStream'
<==== Mercurial data ====>
Branch: vms_4.1

+++ vms/libs/common/src/core/dataprovider/live_stream_params.h
@@ -12,7 +12,7 @@ struct QnLiveStreamParams
 
     Qn::StreamQuality quality = Qn::StreamQuality::undefined;
     float fps = kFpsNotInitialized;
-    int bitrateKbps = 0;
+    float bitrateKbps = 0;
     QSize resolution;
     QString codec;
